Pull progress from Assignment 3 to main

+++ src/watchdog.c
@@ -1,73 +1,91 @@
 #include "watchdog.h"
-#include "constants.h"
-
-#include <stdio.h>  
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-
-#include <sys/types.h>
-#include <sys/shm.h>
-#include <sys/mman.h>
-
-#include <fcntl.h>
-#include <semaphore.h>
-#include <signal.h>
-
-
+#include "common.h"
 
 /* Global variables */
 pid_t server_pid;
-pid_t window_pid;
+pid_t interface_pid;
 pid_t km_pid;
 pid_t drone_pid;
+pid_t obstacles_pid;
+pid_t targets_pid;
+
 pid_t wd_pid;
+pid_t logger_pid;
 
 /* Counters for components*/
 int cnt_server;
 int cnt_window;
 int cnt_km;
 int cnt_drone;
+int cnt_obstacles;
+int cnt_targets;
+
+int cnt_logger;
+
+char logfile[80]; // path to logfile
+char msg[1024];
 
 
 
 void signal_handler(int signo, siginfo_t *siginfo, void *context)
 {
-    printf("Received signal number: %d\n from ", signo);
+    sprintf(msg, "Received signal number: %d\n from ", signo);
+    log_msg(logfile, WD, msg);
     if (signo == SIGINT)
-    {
-        send_sigint_to_all();
-    } 
+        send_sigint_to_all(); 
     if (signo == SIGUSR2)
     {
         // CHECK THE PID OF SENDER AND SET COUNTERS TO ZERO
         if (siginfo->si_pid == server_pid)
         {
-            printf("Server has sent SIGUSR2 \n\n");
+            sprintf(msg, "Received SIGUSR2 from Server\n");
+            log_msg(logfile, WD, msg);
             cnt_server = 0;
         }
-        if (siginfo->si_pid == window_pid)
+        if (siginfo->si_pid == interface_pid)
         {
-            printf("Window has sent SIGUSR2 \n\n");
+            sprintf(msg, "Received SIGUSR2 from Interface \n");
+            log_msg(logfile, WD, msg);
             cnt_window = 0;
         }
         if (siginfo->si_pid == km_pid)
         {
-            printf("Keymanager has sent SIGUSR2 \n\n");
+            sprintf(msg, "Received SIGUSR2 from Key Manager \n");
+            log_msg(logfile, WD, msg);
             cnt_km = 0;
         }
         if (siginfo->si_pid == drone_pid)
         {
-            printf("Drone has sent SIGUSR2 \n\n");
+            sprintf(msg, "Received SIGUSR2 from Drone\n");
+            log_msg(logfile, WD, msg);
             cnt_drone = 0;
         }
+        if (siginfo->si_pid == logger_pid)
+        {
+            sprintf(msg, "Received SIGUSR2 from Logger \n");
+            log_msg(logfile, WD, msg);
+            cnt_logger = 0;
+        }
+        if (siginfo->si_pid == obstacles_pid)
+        {
+            sprintf(msg, "Received SIGUSR2 from Obstacles \n");
+            log_msg(logfile, WD, msg);
+            cnt_obstacles = 0;
+        }
+        if (siginfo->si_pid == targets_pid)
+        {
+            sprintf(msg, "Received SIGUSR2 from Targets\n");
+            log_msg(logfile, WD, msg);
+            cnt_targets = 0;
+        }
     }
-
  }
 
 
 int main(int argc, char* argv[])
 {
+    get_args(argc, argv);
+
     /* Sigaction */
     struct sigaction sa;
     sa.sa_sigaction = signal_handler;
@@ -77,54 +95,79 @@ int main(int argc, char* argv[])
 
     /* Get PIDs of processes*/
     server_pid = 0;
-    window_pid = 0;
+    interface_pid = 0;
     km_pid = 0;
     drone_pid = 0;
+    logger_pid = 0;
+    obstacles_pid = 0;
+    targets_pid = 0;
     wd_pid = getpid();
 
-
     // Get pid of the processes
-    get_pids(&server_pid, &window_pid, &km_pid, &drone_pid);
-    printf("WD PID IS: %d\n", wd_pid);
+    get_pids(&server_pid, &interface_pid, &km_pid, &drone_pid, &obstacles_pid, &targets_pid);
+    sprintf(msg, "PID of WATCHDOG: %d\n", wd_pid);
+    log_msg(logfile, WD, msg);
+
+    // decide on mode
+    char mode[80];
+    read_args_from_file(CONFIG_PATH, mode, NULL);
 
     /*TODO: remove hardcoded values in WATCHDOG */
     cnt_server = 0;
     cnt_window = 0;
     cnt_km = 0;
     cnt_drone = 0;
+    cnt_logger = 0;
+    cnt_obstacles = 0;
+    cnt_targets = 0;
 
-    while(1)
+    while (1)
     {
-        // increment counter
-        cnt_server++;
-        cnt_window++;
-        cnt_km++;
-        cnt_drone++;
-
-        /* Monitor health of all of the processes */
-        kill(server_pid,SIGUSR1);
-        usleep(500);
-        kill(window_pid, SIGUSR1);
-        usleep(500);
-        kill(km_pid, SIGUSR1);
-        usleep(500);
-        kill(drone_pid, SIGUSR1);
-        usleep(500);
-
-
-        // If any of the processess does not respond in given timeframe, close them all
-        if (cnt_server > THRESHOLD || cnt_window > THRESHOLD || cnt_km > THRESHOLD || cnt_drone > THRESHOLD)
+        // do diagnostics only for standalone
+        if (strcmp(mode, "standalone") == 0)
         {
-            send_sigint_to_all();
+            // increment counter
+            cnt_server++;
+            cnt_window++;
+            cnt_km++;
+            cnt_drone++;
+            // FIXME: Unknown problems with signal handling in obstacles and targets
+            // cnt_obstacles++;
+            // cnt_targets++;
+
+            /* Monitor health of all of the processes */
+            kill(server_pid, SIGUSR1);
+            usleep(50);
+            kill(interface_pid, SIGUSR1);
+            usleep(50);
+            kill(km_pid, SIGUSR1);
+            usleep(50);
+            kill(drone_pid, SIGUSR1);
+            usleep(50);
+            kill(targets_pid, SIGUSR1);
+            usleep(50);
+            kill(obstacles_pid, SIGUSR1);
+            usleep(50);
+
+            // If any of the processess does not respond in given timeframe, close them all
+            if (cnt_server > THRESHOLD || cnt_window > THRESHOLD || cnt_km > THRESHOLD || cnt_drone > THRESHOLD ||
+                cnt_targets > THRESHOLD || cnt_obstacles > THRESHOLD)
+            {
+                sprintf(msg, "One of the counters has went through threshold, Threshold value: %d, values - server: %d window : % d, key manager : % d, drone : % d, targets : % d, obstacles : % d",
+                        THRESHOLD, cnt_server, cnt_window, cnt_km, cnt_drone, cnt_targets, cnt_obstacles);
+                log_msg(logfile, WD, msg);
+                send_sigint_to_all();
+            }
         }
-        usleep(100000);
+        usleep(1000000);
     }
 
     return 0;
 }
 
 
-int get_pids(pid_t *server_pid, pid_t *window_pid, pid_t *km_pid, pid_t *drone_pid)
+int get_pids(pid_t *server_pid, pid_t *interface_pid, pid_t *km_pid, 
+                pid_t *drone_pid, pid_t *obstacles_pid, pid_t *targets_pid)
 {
     sem_t *sem_wd_1 = sem_open(SEM_WD_1, 0);
     sem_t *sem_wd_2 = sem_open(SEM_WD_2, 0);
@@ -136,7 +179,7 @@ int get_pids(pid_t *server_pid, pid_t *window_pid, pid_t *km_pid, pid_t *drone_p
 
     // give initial 'signal' to processes that WD is ready to get the pids
     sem_post(sem_wd_1);
-    for (int i = 0; i < 4; i++)
+    for (int i = 0; i < 6; i++)
     {   
         // wait for a signal from processes that you can acccess pids
         sem_wait(sem_wd_2);
@@ -145,27 +188,44 @@ int get_pids(pid_t *server_pid, pid_t *window_pid, pid_t *km_pid, pid_t *drone_p
         int symbol = 0;
         pid_t pid_temp;
         sscanf(ptr_wd, "%i %i", &symbol, &pid_temp);
+        sprintf(msg,"current pid %d  symbol %i \n", pid_temp, symbol);
+        log_msg(logfile, WD, msg);
         switch (symbol)
         {
         case SERVER_SYM:
             *server_pid = pid_temp;
-            printf("Server PID SET\n");
+            sprintf(msg, "Server PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);
             break;
         case WINDOW_SYM:
-            *window_pid = pid_temp;
-            printf("WINDOW PID SET\n");
+            *interface_pid = pid_temp;
+            sprintf(msg,"Interface PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);
             break;
         case KM_SYM:
             *km_pid = pid_temp;
-            printf("KEYMANAGER PID SET\n");
+            sprintf(msg,"Key Manager PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);
             break;
         case DRONE_SYM:
             *drone_pid = pid_temp;
-            printf("DRONE PID SET\n");
+            sprintf(msg,"Drone PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);
+            break;
+        case OBSTACLES_SYM:
+            *obstacles_pid = pid_temp;
+            sprintf(msg,"Obstacles PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);
+            break;
+        case TARGETS_SYM:
+            *targets_pid = pid_temp;
+            sprintf(msg,"Targets PID SET: %d", pid_temp);
+            log_msg(logfile, WD, msg);    
             break;
         default:
-            perror("Wrong process symbol!");
-            exit(1);
+            sprintf(msg, "Wrong process symbol!: %d ", symbol);
+            log_err(logfile, WD, msg);
+            // exit(1);
         }
 
         // clear memory to make sure we got everything we need
@@ -175,11 +235,12 @@ int get_pids(pid_t *server_pid, pid_t *window_pid, pid_t *km_pid, pid_t *drone_p
         sem_post(sem_wd_3);
     }
     
-    if (*server_pid != 0 && *window_pid != 0 && *km_pid != 0 && *drone_pid != 0)
+    if (*server_pid != 0 && *interface_pid != 0 && *km_pid != 0 && *drone_pid != 0 
+        && *obstacles_pid != 0 && *targets_pid != 0) 
     {
         // Got all pids
-        printf("Those are obtained pids: \n");
-        printf(" SERVER : %i \n WINDOW : %i \n KM : %i \n DRONE : %i \n", *server_pid, *window_pid, *km_pid, *drone_pid);
+        sprintf(msg, "Got all pids");
+        log_msg(logfile, WD, msg);
     }
     else
     {
@@ -199,11 +260,19 @@ int get_pids(pid_t *server_pid, pid_t *window_pid, pid_t *km_pid, pid_t *drone_p
 
 void send_sigint_to_all()
 {
-        kill(window_pid, SIGINT);
-        kill(km_pid, SIGINT);
-        kill(drone_pid, SIGINT);
-        kill(server_pid, SIGINT);
-        printf("WD sent SIGINT to all processes, exiting...\n");
-        exit(1);
+    kill(interface_pid, SIGINT);
+    kill(km_pid, SIGINT);
+    kill(drone_pid, SIGINT);
+    kill(server_pid, SIGINT);
+    kill(obstacles_pid, SIGINT);
+    kill(targets_pid, SIGINT);
+    // kill(logger_pid, SIGINT);
+    sprintf(msg, "WD sent SIGINT to all processes, exiting...");
+    log_msg(logfile, WD, msg);
+    exit(1);
 }
 
+void get_args(int argc, char *argv[])
+{
+    sscanf(argv[1], "%s", logfile);
+}
